feat: refactored names of signature and variable names
